backend/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
mysqlclient==2.2.4
python-dotenv==1.0.1
Pillow==10.2.0 